+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 Added:
 
 - Configuration option per toast type for showing content in toasts
+- Configuration option to hide `nick` and `mode` change server messages
+- Context menu item to server buffers to mark all messages on the server as read
 
 Thanks:
 
 - Bug reports: @darienm, @mercster
-- Feature requests: @rossburton
+- Feature requests: @rossburton, death916
 
 # 2025.6 (2025-06-14)